rec_row.c:157:26: error: implicit conversion loses integer precision: 'u_int' (aka 'unsigned int') to 'uint8_t' (aka 'unsigned char') [-Werror,-Wconversion]

+++ src/include/reconcile.h
@@ -254,7 +254,7 @@ struct __wt_reconcile {
 
 /* Don't increase key prefix-compression unless there's a significant gain. */
 #define WT_KEY_PREFIX_PREVIOUS_MINIMUM 10
-    u_int key_pfx_last; /* Last prefix compression */
+    uint8_t key_pfx_last; /* Last prefix compression */
 
     bool key_pfx_compress;      /* If can prefix-compress next key */
     bool key_pfx_compress_conf; /* If prefix compression configured */
